+++ index.d.ts
@@ -147,5 +147,6 @@ export class AsYouType {
   country: CountryCode | undefined;
   getNumber(): PhoneNumber | undefined;
   getNationalNumber(): string;
-  getTemplate(): string | undefined;
+  getChars(): string;
+  getTemplate(): string;
 }