+++ source/AsYouType.test.js
@@ -9,7 +9,7 @@ class AsYouType extends AsYouType_ {
 
 const USE_NON_GEOGRAPHIC_COUNTRY_CODE = false
 
-describe('as you type', () => {
+describe('AsYouType', () => {
 	it('should use "national_prefix_formatting_rule"', () => {
 		// With national prefix (full).
 		new AsYouType('RU').input('88005553535').should.equal('8 (800) 555-35-35')
@@ -884,13 +884,13 @@ describe('as you type', () => {
 	it('should return PhoneNumber with autocorrected international numbers without leading +', () => {
 		// https://github.com/catamphetamine/libphonenumber-js/issues/316
 		const formatter = new AsYouType('FR')
-		formatter.input('33612902554').should.equal('33612902554')
+		formatter.input('33612902554').should.equal('33 6 12 90 25 54')
 		formatter.getNumber().country.should.equal('FR')
 		formatter.getNumber().nationalNumber.should.equal('612902554')
 		formatter.getNumber().number.should.equal('+33612902554')
 		// Should also strip national prefix.
 		formatter.reset()
-		formatter.input('330612902554').should.equal('330612902554')
+		formatter.input('330612902554').should.equal('33 06 12 90 25 54')
 		formatter.getNumber().country.should.equal('FR')
 		formatter.getNumber().nationalNumber.should.equal('612902554')
 		formatter.getNumber().number.should.equal('+33612902554')
@@ -1044,6 +1044,16 @@ describe('as you type', () => {
 		formatter.input('2').should.equal('14880011 1 2')
 		formatter.input('1').should.equal('14880011 1 21')
 		formatter.input('3').should.equal('14880011 1 213')
+	})
+
+	it('should return the phone number characters entered by the user', () => {
+		const formatter = new AsYouType('RU')
+		formatter.getChars().should.equal('')
+		formatter.input('+123')
+		formatter.getChars().should.equal('+123')
+		formatter.reset()
+		formatter.input('123')
+		formatter.getChars().should.equal('123')
 	})
 })