+++ CHANGELOG.md
@@ -63,6 +63,13 @@ And edit the README:
 
 <!-- (breaking change) Changed `countries` and `country_calling_codes` properties in metadata: now they're not properties but rather elements of an array (`countries` is an array now rather than an object; `countries` is `metadata[0]` and `country_calling_codes` is `metadata[1]`). If you were using a custom-generated metadata then it has to be re-generated for the new version. -->
 
+1.9.3 / 11.11.2020
+==================
+
+* Added `AsYouType.getChars()` method.
+
+* Added formatting of international phone numbers that have been input without a leading `+`.
+
 1.9.2 / 08.11.2020
 ==================